chore: LAB1 улучшение структуры папок проекта
