+++ backend/package.json
@@ -14,6 +14,6 @@
     "nodemon": "^3.1.9"
   },
   "scripts": {
-    "dev": "nodemon ./index.js"
+    "dev": "nodemon ./src/server.js"
   }
 }
