+++ backend/src/middlewares/errorHandler.js
@@ -1,4 +1,3 @@
-// middlewares/errorHandler.js
 module.exports = (err, req, res, next) => {
     console.error(err);
   
@@ -6,7 +5,6 @@ module.exports = (err, req, res, next) => {
       return res.status(err.statusCode).json({ error: err.message });
     }
   
-    // Если ошибка не кастомная, возвращаем 500
     return res.status(500).json({ error: 'Внутренняя ошибка сервера' });
   };
