+++ backend/src/middlewares/morganLogger.js
@@ -1,0 +1,3 @@
+const morgan = require('morgan');
+
+module.exports = morgan(':method :url :status :response-time ms');
