+++ backend/src/models/index.js
@@ -1,0 +1,10 @@
+const { sequelize } = require('../config/db');
+const User = require('./user.model');
+const Event = require('./event.model');
+
+// Определяем связи
+User.hasMany(Event, { foreignKey: 'createdBy', as: 'events' });
+Event.belongsTo(User, { foreignKey: 'createdBy', as: 'user' });
+
+// Экспортируем модели и sequelize
+module.exports = { sequelize, User, Event };
