+++ backend/src/models/user.model.js
@@ -1,0 +1,26 @@
+const { DataTypes } = require('sequelize');
+const { sequelize } = require('../config/db');
+
+const User = sequelize.define('User', {
+  id: {
+    type: DataTypes.INTEGER,
+    primaryKey: true,
+    autoIncrement: true,
+  },
+  name: {
+    type: DataTypes.STRING,
+    allowNull: false,
+  },
+  email: {
+    type: DataTypes.STRING,
+    allowNull: false,
+    unique: true,
+    validate: {
+      isEmail: true,
+    },
+  },
+}, {
+  timestamps: true,
+});
+
+module.exports = User;
