+++ backend/src/routes/event.routes.js
@@ -1,0 +1,12 @@
+const express = require('express');
+const eventController = require('../controllers/event.controller');
+
+const router = express.Router();
+
+router.get('/', eventController.getAllEvents);
+router.get('/:id', eventController.getEventById);
+router.post('/', eventController.createEvent);
+router.put('/:id', eventController.updateEvent);
+router.delete('/:id', eventController.deleteEvent);
+
+module.exports = router;
