+++ backend/src/routes/user.routes.js
@@ -1,0 +1,9 @@
+const express = require('express');
+const userController = require('../controllers/user.controller');
+
+const router = express.Router();
+
+router.post('/', userController.createUser);
+router.get('/:id', userController.getUserById);
+
+module.exports = router;
